Simple Smart Home App
This small smart home application allows you to on and off your light,
which is connected to Raspberry Pi.

Using the IBM Bluemix Speech-to-Text  SDK on iOS Application to detect
and convert “turn on” and “turn off“ commands,

NodeJS on Server Side to send the commands to Raspberry Pi,

NodeJS on Raspberry Pi, to get and process the commands,

Python on Raspberry Pi, to manage the PINs on Raspberry Pi to turn on
or turn off lights.